Change lint plugin to ALE, auto-fix format on buffer save
While I loved StandardJS when I was first introduced to it, I'm not a
huge fan of having my linter tell me how to format my code when there
are tools like Prettier that will just format things for me.

This commit removes Standard as my global default for JS linting and
formatting in preference for Prettier and react-app ESLint config. To
maintain proper linting for my older projects that use Standard, I check
to see if standard exists as a local binary.

Other minor changes:

- remove unnecessary carriage returns from certain key mappings